reorg(hulls) split out hullGeom2 and hullGeom3
