+++ src/operations/hulls/hullGeom3.js
@@ -1,0 +1,10 @@
+/*
+ * Create a convex hull of the given geometries (geom3).
+ * @param {...geometries} geometries - list of geom3 geometries
+ * @returns {geom3} new geometry
+ */
+const hullGeom3 = (...geometries) => {
+  throw new Error('sorry. hull of 3D geometries is not supported yet')
+}
+
+module.exports = hullGeom3
